+++ config/text/help/fr/intro.yml
@@ -1,5 +1,5 @@
 |
-  L'onglet *Aide* contient des informations sur l'interface utilisateur de 
+  L'onglet **Aide** contient des informations sur l'interface utilisateur de 
   l'application. Pour plus d'informations sur l'interprétation des valeurs et 
   sur la science et les méthodes derrière les données présentées dans 
   Explorateur de valeurs de calcul, veuillez consulter le document
